backend/requirements.txt
# FastAPI core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Configuration
pydantic-settings==2.7.1

# Rate Limiting
slowapi==0.1.9

# Redis
redis[hiredis]==5.2.1

# AI / RAG (Phase 3)
openai>=1.51.0
anthropic>=0.34.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
numpy>=1.24.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
aiosqlite==0.20.0
